unsatisfying but working barycentric solution
