+++ circuits/cpp/barretenberg/cpp/src/barretenberg/honk/sumcheck/polynomials/barycentric_data.test.cpp
@@ -1,5 +1,5 @@
-#include "barretenberg/ecc/curves/bn254/fr.hpp"
 #include "barycentric_data.hpp"
+#include "barretenberg/ecc/curves/bn254/fr.hpp"
 
 #include "barretenberg/numeric/random/engine.hpp"
 #include <gtest/gtest.h>
@@ -13,6 +13,20 @@ using FieldTypes = testing::Types<barretenberg::fr>;
 TYPED_TEST_SUITE(BarycentricDataTests, FieldTypes);
 
 #define BARYCENTIC_DATA_TESTS_TYPE_ALIASES using FF = TypeParam;
+
+/**
+ * @brief Ensure auxilliary arrays (e.g. big_domain) are computed at compile time if possible (i.e. if FF is literal)
+ *
+ */
+TYPED_TEST(BarycentricDataTests, CompileTimeComputation)
+{
+    BARYCENTIC_DATA_TESTS_TYPE_ALIASES
+    const size_t domain_size(2);
+    const size_t num_evals(10);
+    if constexpr (std::is_literal_type_v<FF>) {
+        static_assert(BarycentricData<FF, domain_size, num_evals>::big_domain[5] == 5);
+    }
+}
 
 TYPED_TEST(BarycentricDataTests, Extend)
 {
